+++ app/components/screen/Widget.js
@@ -1,8 +1,12 @@
-import React from 'react';
+import React, { useEffect, useState } from 'react';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 import * as R from 'ramda';
 import WidgetError from './WidgetError';
+
+import {
+  View,
+} from "react-native";
 import {
   AudioImageRecord,
   AudioRecord,
@@ -12,7 +16,6 @@ import {
   Drawing,
   Geolocation,
   MultiSelect,
-  TLMultiSelect,
   Radio,
   Select,
   Slider,
@@ -21,6 +24,8 @@ import {
   TimeRange,
   VisualStimulusResponse,
   RadioPrizes,
+  StackedSlider,
+  StackedRadio,
 } from '../../widgets';
 import TimePicker from '../../widgets/TimeRange/TimePicker';
 import { setSelected } from '../../state/responses/responses.actions';
@@ -43,7 +48,7 @@ const Widget = ({ screen, answer, onChange, applet, isCurrent, isSelected, setSe
         config={screen.valueConstraints}
         onChange={onChange}
         value={screenValue}
-        token={valueType && valueType.includes("token")}
+        token={valueType && valueType.includes('token')}
       />
     );
   }
@@ -61,9 +66,33 @@ const Widget = ({ screen, answer, onChange, applet, isCurrent, isSelected, setSe
       />
     );
   }
+
+  if (screen.inputType === 'stackedRadio') {
+    return (
+      <StackedRadio
+        config={screen.valueConstraints}
+        onChange={onChange}
+        value={answer}
+        token={valueType && valueType.includes("token")}
+      />
+    )
+  }
+
   if (screen.inputType === 'slider') {
     return (
       <Slider
+        config={screen.valueConstraints}
+        appletName={applet.name.en}
+        onChange={onChange}
+        onPress={onPress}
+        onRelease={onRelease}
+        value={answer}
+      />
+    );
+  }
+  if (screen.inputType === 'stackedSlider') {
+    return (
+      <StackedSlider
         config={screen.valueConstraints}
         appletName={applet.name.en}
         onChange={onChange}
@@ -231,7 +260,14 @@ const Widget = ({ screen, answer, onChange, applet, isCurrent, isSelected, setSe
     );
   }
 
-  onContentError();
+  const [oneShot, setOneShot] = useState(false);
+  useEffect(() => {
+    if (onContentError && !oneShot) {
+      setOneShot(true);
+      onContentError();
+    }
+  });
+  
   return <WidgetError />;
 };