+++ app/widgets/AudioRecord/index.js
@@ -1,6 +1,8 @@
+/* eslint-disable radix */
 import React, { Component } from 'react';
 import { View } from 'react-native';
 import PropTypes from 'prop-types';
+import * as R from 'ramda';
 import AudioRecorder from './AudioRecorder';
 
 export class AudioRecord extends Component {
@@ -10,12 +12,14 @@ export class AudioRecord extends Component {
   }
 
   render() {
-    const { value } = this.props;
+    const { value, config } = this.props;
+    const maxLength = parseInt(R.path(['maxValue'], config));
     return (
       <View style={{ paddingTop: 16, paddingBottom: 16 }}>
         <AudioRecorder
           onStop={this.onRecord}
           path={value && value.uri}
+          maxLength={Number.isNaN(maxLength) ? Infinity : maxLength}
         />
       </View>
     );
@@ -31,4 +35,5 @@ AudioRecord.propTypes = {
     uri: PropTypes.string,
   }),
   onChange: PropTypes.func.isRequired,
+  config: PropTypes.object.isRequired,
 };