+++ app/widgets/Camera.js
@@ -203,7 +203,7 @@ export class Camera extends Component {
       const options = {
         mediaType: 'video',
         videoQuality: 'low',
-        durationLimit: 5,
+        durationLimit: 60,
         // saveToPhotos: true,
       };
       if (config.allowLibrary) {